feat: :sparkles: shorter install script
