+++ README.md
@@ -49,7 +49,7 @@
 **Note: You need to upgrade to Deno v1.4.1 or newer in order to use our CLI.**
 
 ```shell script
-deno install -A -f --unstable -n eggs https://x.nest.land/eggs@0.2.3/mod.ts
+deno install -Afq --unstable https://x.nest.land/eggs@0.3.0/eggs.ts
 ```
 
 For more information, see the [documentation](https://docs.nest.land/).
@@ -156,7 +156,7 @@ After `eggs update`:
 ```ts
 import * as colors from "https://deno.land/std@0.58.0/fmt/colors.ts";
 import * as bcrypt from "https://deno.land/x/bcrypt@v0.2.1/mod.ts"
-import * as eggs from "https://x.nest.land/eggs@0.2.3/mod.ts"
+import * as eggs from "https://x.nest.land/eggs@0.3.0/mod.ts"
 import * as http from "https://deno.land/std/http/mod.ts"
 ```
